Refactor common methods out of PushUpdates and into PushUpdatesBase. This will be used as a foundation for other PushUpdatesFor... modules.
